grayscale-app/requirements.txt
Pillow
pika
